bug fix: was always requiring both mongoid & rails
the defined? guards were checking string literals rather than symbols, so were always returning true.  

(for the case of Mongoid this meant that symbolize/mongoid was always being included, which as a side effect caused the symbol ::Mongoid to be defined, which then broke other gems which likewise checked for the existence of that symbol before doing mongo-specific things)